# frontend/openapi/asbl-club-api.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "asbl.club API",
    "description" : "Public read API exposing associations and their public events. Reads are open with CORS for showcase sites. Everything else requires a JWT access token from POST /api/v1/auth/login, sent as a Bearer token.",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
...
    "/api/v1/registrations/{id}/checkout" : {
      "post" : {
        "tags" : [ "Registrations" ],
        "summary" : "Start (or resume) paying for one of my bookings",
        "operationId" : "startCheckout",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Ready to pay",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Checkout"
                }
              }
            }
          },
          "409" : {
            "description" : "Nothing to pay (NOTHING_TO_PAY), not paid in time (BOOKING_EXPIRED), or the association can't receive payments yet (PAYMENTS_DISABLED)",
            "content" : {
              "application/problem+json" : { }
            }
          },
          "502" : {
            "description" : "The payment provider couldn't be reached",
            "content" : {
              "application/problem+json" : { }
            }
          }
        },
        "security" : [ {
          "bearer" : [ ]
        } ]
      }
    },
    "/api/v1/registrations/{id}/cancel" : {
      "post" : {
        "tags" : [ "Registrations" ],
        "summary" : "Cancel one of my paid tickets and be refunded",
        "operationId" : "cancelMyBooking",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Cancelled and refunded",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MyBooking"
                }
              }
            }
          },
          "409" : {
            "description" : "The delay has passed or tickets aren't refundable (CANCELLATION_CLOSED), or not a paid unused ticket of an event still on (NOT_CANCELLABLE)",
            "content" : {
              "application/problem+json" : { }
            }
          },
          "502" : {
            "description" : "The payment provider couldn't be reached",
            "content" : {
              "application/problem+json" : { }
            }
          }
        },
        "security" : [ {
          "bearer" : [ ]
        } ]
      }
    },
...
    "/api/v1/events/{eventId}/registrations" : {
      "post" : {
        "tags" : [ "Registrations" ],
        "summary" : "Book a seat on a public event (any account)",
        "operationId" : "bookPublicEvent",
        "parameters" : [ {
          "name" : "eventId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BookRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Booked; pay next",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MyRegistration"
                }
              }
            }
          },
          "409" : {
            "description" : "Sold out (SOLD_OUT), or the association can't receive payments yet (PAYMENTS_DISABLED)",
            "content" : {
              "application/problem+json" : { }
            }
          }
        },
        "security" : [ {
          "bearer" : [ ]
        } ]
      }
    },
...
    "/api/v1/asbls/{slug}/manage/events/{eventId}/registrations" : {
      "post" : {
        "tags" : [ "Registrations" ],
        "summary" : "Book a seat (members of the association)",
        "operationId" : "bookTicket",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "eventId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BookRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Booked; pay next",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MyRegistration"
                }
              }
            }
          },
          "409" : {
            "description" : "Sold out, or the event isn't open for booking",
            "content" : {
              "application/problem+json" : { }
            }
          }
        },
        "security" : [ {
          "bearer" : [ ]
        } ]
      }
    },
...
    "/api/v1/registrations" : {
      "get" : {
        "tags" : [ "Registrations" ],
        "summary" : "My bookings, soonest event first",
        "operationId" : "listMyBookings",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/MyBooking"
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer" : [ ]
        } ]
      }
    },
    "/api/v1/registrations/{id}" : {
      "get" : {
        "tags" : [ "Registrations" ],
        "summary" : "One of my bookings and its payment status",
        "operationId" : "getMyRegistration",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MyRegistration"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer" : [ ]
        } ]
      }
    },
...
    "/api/v1/events/{eventId}" : {
      "get" : {
        "tags" : [ "Public" ],
        "summary" : "Get a public event with its association, location and tickets",
        "operationId" : "getEvent",
        "parameters" : [ {
          "name" : "eventId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PublicEvent"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/asbls/{slug}" : {
      "get" : {
        "tags" : [ "Public" ],
        "summary" : "Get an association by slug",
        "operationId" : "getAsbl",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AsblResource"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Checkout" : {
        "required" : [ "amount", "clientSecret", "currency", "publishableKey", "stripeAccount" ],
        "type" : "object",
        "properties" : {
          "publishableKey" : {
            "type" : "string"
          },
          "stripeAccount" : {
            "type" : "string",
            "description" : "The association's Stripe account (payments go straight to it)"
          },
          "clientSecret" : {
            "type" : "string"
          },
          "amount" : {
            "type" : "number",
            "description" : "Amount in euros"
          },
          "currency" : {
            "type" : "string"
          }
        }
      },
      "MyBooking" : {
        "required" : [ "amount", "asblName", "asblSlug", "currency", "eventId", "eventTitle", "id", "startsAt", "status", "ticketLabel" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "RESERVED", "PAID", "CONFIRMED", "ATTENDED", "CANCELLED", "REFUNDED", "EXPIRED" ]
          },
          "amount" : {
            "type" : "number",
            "description" : "Amount in euros"
          },
          "currency" : {
            "type" : "string"
          },
          "eventId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "eventTitle" : {
            "type" : "string"
          },
          "startsAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "location" : {
            "type" : "string"
          },
          "asblName" : {
            "type" : "string"
          },
          "asblSlug" : {
            "type" : "string"
          },
          "ticketLabel" : {
            "type" : "string"
          },
          "ticketCode" : {
            "type" : "string",
            "description" : "The code in the ticket's QR code; present once the booking is paid. Whoever has it gets in, so it's only ever shown to the booking's owner."
          },
          "checkedInAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "cancellableUntil" : {
            "type" : "string",
            "description" : "Until when this ticket can still be cancelled and refunded; absent: it can't",
            "format" : "date-time"
          }
        }
      },
...
      "BookRequest" : {
        "required" : [ "ticketCategoryId" ],
        "type" : "object",
        "properties" : {
          "ticketCategoryId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "MyRegistration" : {
        "required" : [ "amount", "asblSlug", "currency", "eventId", "eventTitle", "id", "status", "ticketLabel" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "RESERVED", "PAID", "CONFIRMED", "ATTENDED", "CANCELLED", "REFUNDED", "EXPIRED" ]
          },
          "amount" : {
            "type" : "number",
            "description" : "Amount in euros"
          },
          "currency" : {
            "type" : "string"
          },
          "eventId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "eventTitle" : {
            "type" : "string"
          },
          "ticketLabel" : {
            "type" : "string"
          },
          "asblSlug" : {
            "type" : "string"
          }
        }
      },
...
      "AsblResource" : {
        "required" : [ "denomination", "slug" ],
        "type" : "object",
        "properties" : {
          "slug" : {
            "type" : "string"
          },
          "denomination" : {
            "type" : "string"
          }
        }
      },
...
      "PublicEvent" : {
        "required" : [ "asbl", "id", "startsAt", "tickets", "title" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "startsAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "location" : {
            "type" : "string"
          },
          "asbl" : {
            "$ref" : "#/components/schemas/AsblResource"
          },
          "cancellableUntil" : {
            "type" : "string",
            "description" : "Until when buyers may cancel their ticket and be refunded; absent: not refundable on request",
            "format" : "date-time"
          },
          "tickets" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PublicTicket"
            }
          }
        }
      },
      "PublicTicket" : {
        "required" : [ "id", "label", "price", "remaining" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "label" : {
            "type" : "string"
          },
          "price" : {
            "type" : "number",
            "description" : "Price in euros"
          },
          "remaining" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearer" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}